refactor: move compact utils to include
