+++ lib/jekyll/favicon/utils.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'jekyll/favicon/utils/configuration'
 require 'jekyll/favicon/utils/convert'
 require 'jekyll/favicon/utils/tag'
 
@@ -7,6 +8,7 @@ module Jekyll
   module Favicon
     # Favicon utils functions
     module Utils
+      include Favicon::Utils::Configuration
       include Favicon::Utils::Convert
       include Favicon::Utils::Tag
 
@@ -23,13 +25,6 @@ module Jekyll
       end
 
       class << self
-        def compact(compactable)
-          case compactable
-          when Hash, Array then deep_compact(compactable) || compactable.class[]
-          else compactable
-          end
-        end
-
         def find_all(findable, target)
           findable.each_with_object([]) do |(key, value), memo|
             if key == target
@@ -70,34 +65,6 @@ module Jekyll
           return merged if rest.empty?
 
           merge(merged, *rest)
-        end
-
-        private
-
-        def deep_compact(compactable)
-          case compactable
-          when Hash then compact_hash compactable
-          when Array then compact_array compactable
-          else compactable
-          end
-        end
-
-        def compact_hash(hash)
-          compacted_hash = hash.each_with_object({}) do |(key, value), memo|
-            next unless (compacted = deep_compact(value))
-
-            memo[key] = compacted
-          end
-          compacted_hash.empty? ? nil : compacted_hash.compact
-        end
-
-        def compact_array(array)
-          compacted_array = array.each_with_object([]) do |value, memo|
-            next unless (compacted = deep_compact(value))
-
-            memo << compacted
-          end
-          compacted_array.empty? ? nil : compacted_array.compact
         end
       end
     end
